+++ app/src/main/AndroidManifest.xml
@@ -16,22 +16,22 @@
         <activity android:name=".UI.Developer.AddEmployeeActivity" />
         <activity android:name=".TestHelper.IssueTrackerHelper" />
         <activity android:name=".UI.InfoBarActivity" />
-        <activity android:name=".ShowSaleListActivity" />
-        <activity android:name=".AddSaleActivity" />
-        <activity android:name=".RemoveSaleActivity" />
-        <activity android:name=".RemoveTaskActivity" />
-        <activity android:name=".AddTaskActivity" />
-        <activity android:name=".ShowTaskListActivity" />
-        <activity android:name=".RemoveClientActivity" />
-        <activity android:name=".AddClientActivity" />
-        <activity android:name=".ShowClientListActivity" />
-        <activity android:name=".ManageTaskActivity" />
-        <activity android:name=".SubmitTicketClientActivity" />
-        <activity android:name=".ManageSaleActivity" />
-        <activity android:name=".ManageClientActivity" />
+        <activity android:name=".ClientMenu.ShowSaleListActivity" />
+        <activity android:name=".ClientMenu.AddSaleActivity" />
+        <activity android:name=".ClientMenu.RemoveSaleActivity" />
+        <activity android:name=".ClientMenu.RemoveTaskActivity" />
+        <activity android:name=".ClientMenu.AddTaskActivity" />
+        <activity android:name=".ClientMenu.ShowTaskListActivity" />
+        <activity android:name=".ClientMenu.RemoveClientActivity" />
+        <activity android:name=".ClientMenu.AddClientActivity" />
+        <activity android:name=".ClientMenu.ShowClientListActivity" />
+        <activity android:name=".ClientMenu.ManageTaskActivity" />
+        <activity android:name=".ClientMenu.SubmitTicketClientActivity" />
+        <activity android:name=".ClientMenu.ManageSaleActivity" />
+        <activity android:name=".ClientMenu.ManageClientActivity" />
         <activity android:name=".Employee.ApplicationsActivity" tools:ignore="InnerclassSeparator" />
-        <activity android:name=".ClientMenuActivity" />
-        <activity android:name=".EmployeeMenuActivity" />
+        <activity android:name=".ClientMenu.ClientMenuActivity" />
+        <activity android:name=".ClientMenu.EmployeeMenuActivity" />
         <activity android:name=".UI.Developer.DeveloperMenuActivity" tools:ignore="InnerclassSeparator" />
         <activity android:name=".UI.Developer.IssueTracker.ManageIssueTrackerActivity" tools:ignore="InnerclassSeparator" />
         <activity android:name=".UI.Developer.IssueTracker.AddOrUpdateIssueTrackerActivity" tools:ignore="InnerclassSeparator" />
