--- app/src/main/java/com/example/forcesales/AddSaleActivity.java
@@ -1,13 +1,0 @@
-package com.example.forcesales;
-
-import androidx.appcompat.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class AddSaleActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_add_sale);
-    }
-}
