+++ app/src/main/java/com/example/forcesales/ClientMenu/ManageSaleActivity.java
@@ -1,4 +1,4 @@
-package com.example.forcesales;
+package com.example.forcesales.ClientMenu;
 
 import android.content.Intent;
 import androidx.appcompat.app.AppCompatActivity;
@@ -10,6 +10,7 @@ import android.widget.Toast;
 
 import com.example.forcesales.Data.Client.Client;
 import com.example.forcesales.Data.Sale.Sales;
+import com.example.forcesales.R;
 
 import java.util.ArrayList;
