+++ app/src/main/java/com/example/forcesales/ClientMenu/ManageTaskActivity.java
@@ -1,4 +1,4 @@
-package com.example.forcesales;
+package com.example.forcesales.ClientMenu;
 
 import android.content.Intent;
 import androidx.appcompat.app.AppCompatActivity;
@@ -7,8 +7,10 @@ import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
+
 import com.example.forcesales.Data.Client.Client;
 import com.example.forcesales.Data.Tasks.Task;
+import com.example.forcesales.R;
 
 import java.util.ArrayList;
