+++ app/src/main/java/com/example/forcesales/ClientMenu/RemoveClientActivity.java
@@ -1,41 +1,37 @@
-package com.example.forcesales;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
+package com.example.forcesales.ClientMenu;
 
 import android.content.Intent;
-import android.os.Bundle;
 import android.os.Parcelable;
+import androidx.appcompat.app.AppCompatActivity;
+import android.os.Bundle;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.Log;
+import com.example.forcesales.R;
 
-import com.example.forcesales.Data.Sale.Sale;
-import com.example.forcesales.Data.Sale.Sales;
-import com.example.forcesales.Data.Tasks.TaskList;
-import com.example.forcesales.RecycleViewItems.RemoveSaleAdapter;
-import com.example.forcesales.RecycleViewItems.RemoveTaskAdapter;
+import com.example.forcesales.Data.Client.ClientList;
+import com.example.forcesales.RecycleViewItems.RemoveClientAdapter;
 
-import java.util.ArrayList;
-import java.util.List;
+public class RemoveClientActivity extends AppCompatActivity {
 
-public class RemoveSaleActivity extends AppCompatActivity {
     private RecyclerView mRecyclerView;
     private RecyclerView.Adapter mAdapter;
     private RecyclerView.LayoutManager mLayoutManager;
-    Sales _Sale;
+    private ClientList _List;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_recycleview_list);
 
-        _Sale = getIntent().getParcelableExtra("SALES_LIST");
-        ArrayList<Sale> _List = _Sale.getSalesList();
+        getSupportActionBar().setTitle("Remove Clients");
+
+        _List =  getIntent().getParcelableExtra("ACCOUNT_LIST");
 
         mRecyclerView = findViewById(R.id.abstract_recycleview_list);
         mRecyclerView.setHasFixedSize(true);
         mLayoutManager = new LinearLayoutManager(this);
-        mAdapter = new RemoveSaleAdapter(_List);
+        mAdapter = new RemoveClientAdapter(_List);
 
         mRecyclerView.setLayoutManager(mLayoutManager);
         mRecyclerView.setAdapter(mAdapter);
@@ -45,11 +41,11 @@ public class RemoveSaleActivity extends AppCompatActivity {
     @Override
     public void onBackPressed() {
 
-        Log.d("APP", "onBackPressed() Override called it RemoveSaleActivity");
+        Log.d("APP", "onBackPressed() Override called it RemoveClientActivity");
+
         Intent result = new Intent();
-        result.putExtra("SALES_LIST", (Parcelable) _Sale);
+        result.putExtra("ACCOUNT_LIST", (Parcelable) _List);
         setResult(RESULT_OK, result);
         finish();
     }
-
 }
