+++ app/src/main/java/com/example/forcesales/ClientMenu/RemoveSaleActivity.java
@@ -1,36 +1,42 @@
-package com.example.forcesales;
+package com.example.forcesales.ClientMenu;
 
-import android.content.Intent;
-import android.os.Parcelable;
 import androidx.appcompat.app.AppCompatActivity;
-import android.os.Bundle;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.os.Parcelable;
 import android.util.Log;
 
-import com.example.forcesales.Data.Client.ClientList;
-import com.example.forcesales.RecycleViewItems.RemoveClientAdapter;
+import com.example.forcesales.Data.Sale.Sale;
+import com.example.forcesales.Data.Sale.Sales;
+import com.example.forcesales.Data.Tasks.TaskList;
+import com.example.forcesales.R;
+import com.example.forcesales.RecycleViewItems.RemoveSaleAdapter;
+import com.example.forcesales.RecycleViewItems.RemoveTaskAdapter;
 
-public class RemoveClientActivity extends AppCompatActivity {
+import java.util.ArrayList;
+import java.util.List;
 
+public class RemoveSaleActivity extends AppCompatActivity {
     private RecyclerView mRecyclerView;
     private RecyclerView.Adapter mAdapter;
     private RecyclerView.LayoutManager mLayoutManager;
-    private ClientList _List;
+    Sales _Sale;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_recycleview_list);
 
-        getSupportActionBar().setTitle("Remove Clients");
-
-        _List =  getIntent().getParcelableExtra("ACCOUNT_LIST");
+        _Sale = getIntent().getParcelableExtra("SALES_LIST");
+        ArrayList<Sale> _List = _Sale.getSalesList();
 
         mRecyclerView = findViewById(R.id.abstract_recycleview_list);
         mRecyclerView.setHasFixedSize(true);
         mLayoutManager = new LinearLayoutManager(this);
-        mAdapter = new RemoveClientAdapter(_List);
+        mAdapter = new RemoveSaleAdapter(_List);
 
         mRecyclerView.setLayoutManager(mLayoutManager);
         mRecyclerView.setAdapter(mAdapter);
@@ -40,11 +46,11 @@ public class RemoveClientActivity extends AppCompatActivity {
     @Override
     public void onBackPressed() {
 
-        Log.d("APP", "onBackPressed() Override called it RemoveClientActivity");
-
+        Log.d("APP", "onBackPressed() Override called it RemoveSaleActivity");
         Intent result = new Intent();
-        result.putExtra("ACCOUNT_LIST", (Parcelable) _List);
+        result.putExtra("SALES_LIST", (Parcelable) _Sale);
         setResult(RESULT_OK, result);
         finish();
     }
+
 }
