+++ app/src/main/java/com/example/forcesales/ClientMenu/SubmitTicketClientActivity.java
@@ -1,0 +1,15 @@
+package com.example.forcesales.ClientMenu;
+
+import androidx.appcompat.app.AppCompatActivity;
+import android.os.Bundle;
+
+import com.example.forcesales.R;
+
+public class SubmitTicketClientActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_submit_ticket_client);
+    }
+}
