+++ app/src/main/res/layout/activity_add_sale.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddSaleActivity">
+    tools:context=".ClientMenu.AddSaleActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
