+++ app/src/main/res/layout/activity_client_menu.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ClientMenuActivity">
+    tools:context=".ClientMenu.ClientMenuActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
