+++ app/src/main/res/layout/activity_employee_menu.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EmployeeMenuActivity">
+    tools:context=".ClientMenu.EmployeeMenuActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
