+++ app/src/main/res/layout/activity_manage_client.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ManageClientActivity">
+    tools:context=".ClientMenu.ManageClientActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
