+++ app/src/main/res/layout/activity_manage_issue_tracker.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ManageTaskActivity">
+    tools:context=".ClientMenu.ManageTaskActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
