+++ app/src/main/res/layout/activity_manage_sale.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ManageSaleActivity">
+    tools:context=".ClientMenu.ManageSaleActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
