+++ app/src/main/res/layout/activity_submit_ticket_client.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddClientActivity">
+    tools:context=".ClientMenu.AddClientActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
